+++ lib/recurly/account_acquisition.php
@@ -1,0 +1,54 @@
+<?php
+
+class Recurly_AccountAcquisition extends Recurly_Resource
+{
+  protected static $_writeableAttributes;
+
+  public static function init() {
+    Recurly_AccountAcquisition::$_writeableAttributes = array(
+      'cost_in_cents', 'currency', 'channel', 'subchannel', 'campaign',
+    );
+  }
+
+  public static function get($accountCode, $client = null) {
+    return Recurly_Base::_get(Recurly_AccountAcquisition::uriForAccountAcquisition($accountCode), $client);
+  }
+
+  public static function deleteForAccount($accountCode, $client = null) {
+    return Recurly_Base::_delete(Recurly_AccountAcquisition::uriForAccountAcquisition($accountCode), $client);
+  }
+
+  protected static function uriForAccountAcquisition($accountCode) {
+    return Recurly_Client::PATH_ACCOUNTS . '/' . rawurlencode($accountCode) . Recurly_Client::PATH_ACQUISITION;
+  }
+
+  public function create() {
+    $this->update();
+  }
+
+  public function update() {
+    $this->_save(Recurly_Client::PUT, $this->uri());
+  }
+
+  public function delete() {
+    return Recurly_Base::_delete($this->uri(), $this->_client);
+  }
+
+  protected function uri() {
+    if (!empty($this->_href))
+      return $this->getHref();
+    else if (!empty($this->account_code))
+      return Recurly_AccountAcquisition::uriForAccountAcquisition($this->account_code);
+    else
+      throw new Recurly_Error("'account_code' not specified.");
+  }
+
+  protected function getNodeName() {
+    return 'account_acquisition';
+  }
+  protected function getWriteableAttributes() {
+    return Recurly_AccountAcquisition::$_writeableAttributes;
+  }
+}
+
+Recurly_AccountAcquisition::init();