+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added support for `original_transaction` to `Recurly_Transaction` [#238](https://github.com/recurly/recurly-client-php/pull/238)
 * Added `Recurly_AccountBalance` [#239](https://github.com/recurly/recurly-client-php/pull/239)
+* Add support for setting account acquisition
 
 ## Version 2.5.3 (unreleased)